Removed uintptr from mocknetns, tests compile
Forgot to remove uintptr from mocknetns

+++ netns/mocknetns.go
@@ -22,25 +22,25 @@ func NewMock(failMethod int, failMessage string) *MockNetns {
 	}
 }
 
-func (f *MockNetns) Get() (uintptr, error) {
+func (f *MockNetns) Get() (int, error) {
 	if f.failMethod == 1 {
 		return 0, newErrorMockNetns(f.failMessage)
 	}
 	return 1, nil
 }
-func (f *MockNetns) GetFromName(name string) (uintptr, error) {
+func (f *MockNetns) GetFromName(name string) (int, error) {
 	if f.failMethod == 2 {
 		return 0, newErrorMockNetns(f.failMessage)
 	}
 	return 1, nil
 }
-func (f *MockNetns) Set(handle uintptr) error {
+func (f *MockNetns) Set(handle int) error {
 	if f.failMethod == 3 {
 		return newErrorMockNetns(f.failMessage)
 	}
 	return nil
 }
-func (f *MockNetns) NewNamed(name string) (uintptr, error) {
+func (f *MockNetns) NewNamed(name string) (int, error) {
 	if f.failMethod == 4 {
 		return 0, newErrorMockNetns(f.failMessage)
 	}
